requirements.txt
accelerate==0.33.0
aiofiles==24.1.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
albucore==0.0.20
albumentations==1.4.21
altair==5.5.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
backoff==2.2.1
banks==2.1.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cryptography==44.0.2
ctransformers==0.2.27
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
emoji==2.14.1
eval_type_backport==0.2.0
exceptiongroup==1.2.2
executing==2.0.1
faiss-cpu==1.8.0.post1
fastapi==0.115.12
fastjsonschema==2.20.0
filelock==3.15.4
filetype==1.2.0
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
gpt4all==2.7.0
greenlet==3.0.3
griffe==1.7.1
gritql==0.2.0
gTTS==2.5.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.3
idna==3.7
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib_metadata==8.2.0
importlib_resources==6.4.4
ipykernel==6.29.5
ipython==8.18.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.9.0
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.7
langchain==0.3.22
langchain-community==0.3.20
langchain-core==0.3.49
langchain-text-splitters==0.3.7
langdetect==1.0.9
langserve==0.3.1
langsmith==0.3.19
litellm==1.65.0
llama-cloud==0.1.17
llama-cloud-services==0.6.9
llama-index==0.12.26
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.27
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.9
llama-index-llms-litellm==0.3.1
llama-index-llms-openai==0.3.29
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.4.post1
llama_cpp_python==0.3.8
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
narwhals==1.32.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.6.0.22
nvidia-cuda-runtime-cu12==12.6.37
olefile==0.47
openai==1.69.0
orjson==3.10.6
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pillow==10.2.0
platformdirs==4.3.7
proglog==0.1.10
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==5.29.4
psutil==6.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
pydeck==0.9.1
pygame==2.6.0
Pygments==2.18.0
pyparsing==3.1.4
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-iso639==2025.2.18
python-json-logger==2.0.7
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
RapidFuzz==3.12.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.19.1
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.13.1
Send2Trash==1.8.3
sentence-transformers==3.0.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.46.1
streamlit==1.44.0
streamlit-chat==0.1.1
striprtf==0.0.26
sympy==1.13.1
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.9.0
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.13.2
torch==2.4.0+cu118
torchaudio==2.4.0+cu118
torchvision==0.19.0+cu118
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
transformers==4.43.3
typer==0.15.2
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.1
unstructured==0.17.2
unstructured-client==0.32.0
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.34.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
wrapt==1.17.2
yarl==1.9.4
zipp==3.19.2
zstandard==0.23.0
